requirements.txt
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
chromedriver-binary==91.0.4472.101.0
click==8.0.1
cycler==0.10.0
filelock==3.0.12
fuzzywuzzy==0.18.0
gensim==4.0.1
huggingface-hub==0.0.12
idna==2.10
isort==5.8.0
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.2
mccabe==0.6.1
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pep8==1.7.1
Pillow==8.2.0
progressbar2==3.53.1
pycodestyle==2.7.0
pylint==2.8.3
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
six==1.16.0
sklearn==0.0
smart-open==5.1.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
transformers==4.8.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wrapt==1.12.1
